use a unified function for checking reserved suffixes
